Add optimized and looped scene 1 bgm
Trimmed away the front silent part to reduce audible silence when
looping, trimmed away the end so that after playing the last note, the
audio ends, making it loopable so that when playing from the start
again, audio sounds continuous.

Unfortunately, this is not easily loopable because instrumental
effects like reverberation and sustain will not carry over, which will
cause the start to sound abrupt when coming from the end.

Optimized the audio so that it is "streaming" instead of "decompress
on load" because decompressing it will take up space. Quality reduced
to reduce size of the clip used in Unity, though for most people they
won't hear the quality loss.